Edit income condition py file
