+++ policyengine_uk/parameters/gov/dwp/extended_childcare_entitlement/income/income_limit.yaml
@@ -6,7 +6,5 @@ metadata:
   reference:
     - title: The Childcare Payments (Eligibility) Regulations 2015 - Regulation 15
       href: https://www.legislation.gov.uk/uksi/2015/448/regulation/15#commentary-key-e27c923eee152accd495af8425536e29
-    - title: Extended childcare entitlement guidance
-      href: https://www.gov.uk/check-eligible-free-childcare-if-youre-working?step-by-step-nav=f517cd57-3c18-4bb9-aa8b-1b907e279bf9
 values:
   2015-01-01: 100_000
