--- policyengine_uk/parameters/gov/dwp/extended_childcare_entitlement/income_thresholds.yaml
@@ -1,26 +1,0 @@
-description: The Department for Education requires minimum earnings thresholds for extended childcare entitlement eligibility based on the age of parents.
-metadata:
-  period: year
-  type: single_amount 
-  amount_unit: currency-GBP
-  threshold_unit: year
-  label: Extended childcare entitlement minimum earnings threshold
-  reference:
-    - title: Childcare Payments Act 2014 - c. 28 - section 9 (national minimum wage for 8 hours a week)
-      href: https://www.legislation.gov.uk/ukpga/2014/28/notes/division/6/3/4
-    - title: Extended childcare entitlement guidance
-      href: https://www.gov.uk/check-eligible-free-childcare-if-youre-working?step-by-step-nav=f517cd57-3c18-4bb9-aa8b-1b907e279bf9
-
-brackets:
-  - threshold:
-      2015-01-01: 0
-    amount:
-      2015-01-01: 1_331
-  - threshold:
-      2015-01-01: 18
-    amount:
-      2015-01-01: 1_788
-  - threshold:
-      2015-01-01: 21
-    amount:
-      2015-01-01: 2_379
