+++ policyengine_uk/tests/policy/baseline/gov/dwp/extended_childcare_entitlement/extended_childcare_entitlement_income_condition.yaml
@@ -2,108 +2,47 @@
   period: 2025
   input:
     age: 17
-    total_income: 5400  # Above £1,331 * 4 = £5,324 yearly threshold
-    private_pension_income: 0
-    savings_interest_income: 0
-    dividend_income: 0
-    property_income: 0
+    employment_income: 10400  
   output:
-    extended_childcare_entitlement_income_eligible: true
-
-- name: Just turning 18 with yearly threshold
-  period: 2025
-  input:
-    age: 18
-    total_income: 7200  # Above £1,788 * 4 = £7,152 yearly threshold
-    private_pension_income: 0
-    savings_interest_income: 0
-    dividend_income: 0
-    property_income: 0
-  output:
-    extended_childcare_entitlement_income_eligible: true
+    extended_childcare_entitlement_meets_income_requirements: true
 
 - name: Age 20 meeting 18-20 yearly threshold
   period: 2025
   input:
     age: 20
-    total_income: 22000 
-    private_pension_income: 10
-    savings_interest_income: 0
-    dividend_income: 0
-    property_income: 0
+    employment_income: 22000 
   output:
-    extended_childcare_entitlement_income_eligible: true
+    extended_childcare_entitlement_meets_income_requirements: true
 
 - name: Age 21+ meeting higher yearly threshold
   period: 2025
   input:
     age: 25
-    total_income: 29600  # Above £2,379 * 4 = £9,516 yearly threshold
-    private_pension_income: 0
-    savings_interest_income: 0
-    dividend_income: 0
-    property_income: 0
+    employment_income: 29600  # Above £2,379 * 4 = £9,516 yearly threshold
   output:
-    extended_childcare_entitlement_income_eligible: true
-
-- name: Below yearly threshold after investment deductions
-  period: 2025
-  input:
-    age: 30
-    total_income: 12000
-    private_pension_income: 2000
-    savings_interest_income: 1000
-    dividend_income: 500  # Leaving £8,500 yearly < £9,516 threshold
-    property_income: 0
-  output:
-    extended_childcare_entitlement_income_eligible: false
-
+    extended_childcare_entitlement_meets_income_requirements: true
 
 - name: Under 18 below yearly threshold
   period: 2025
   input:
     age: 16
-    total_income: 5000  # Below £5,324 yearly threshold
-    private_pension_income: 0
-    savings_interest_income: 0
-    dividend_income: 0
-    property_income: 0
+    employment_income: 5000  # Below £5,324 yearly threshold
   output:
-    extended_childcare_entitlement_income_eligible: false
+    extended_childcare_entitlement_meets_income_requirements: false
 
-- name: Age 18-20 with investment reducing below yearly threshold
+- name: Age 18-20 - eligible
   period: 2025
   input:
     age: 19
-    total_income: 8000
-    private_pension_income: 0
-    savings_interest_income: 1000  # Leaving £7,000 < £7,152 yearly threshold
-    dividend_income: 0
-    property_income: 0
+    employment_income: 8000
   output:
-    extended_childcare_entitlement_income_eligible: false
-
-- name: All investment income
-  period: 2025
-  input:
-    age: 25
-    total_income: 15000
-    private_pension_income: 5000
-    savings_interest_income: 4000
-    dividend_income: 3000
-    property_income: 3000  # All income is investment = £0 eligible
-  output:
-    extended_childcare_entitlement_income_eligible: false
+    extended_childcare_entitlement_meets_income_requirements: false
 
 - name: At maximum threshold
   period: 2025
   input:
     age: 35
-    total_income: 1000000  # At assumed maximum threshold
+    employment_income: 1000000  # At assumed maximum threshold
     adjusted_net_income: 1000000
-    private_pension_income: 0
-    savings_interest_income: 0
-    dividend_income: 0
-    property_income: 0
   output:
-    extended_childcare_entitlement_income_eligible: false
+    extended_childcare_entitlement_meets_income_requirements: false
